Refined CPUAccelerator implementation to support new multiprocessor classes.
